refactor(components): refactor split utils to each function
refactor split utils to each function. this is for InputField refactoring

close AUTH-AUTH-2787
