+++ components/src/utils/humanizeLabwareType.ts
@@ -1,0 +1,3 @@
+export const humanizeLabwareType = (labwareType: string): string => {
+  return labwareType.replace(/-|_/g, ' ')
+}
